Enable dirty flag
For files that don't support as-you-type but are loaded by live HTML
documents (e.g. JS files), we want to show a dirty indicator on the
live development icon when they have unsaved changes, so the user knows
s/he needs to save in order to have the page reload.